+++ samples/mssql/ServerSideBlazorApp/Pages/Customers.razor.cs
@@ -4,78 +4,133 @@ using Microsoft.AspNetCore.Components;
 using ServerSideBlazorApp.Models;
 using System.Collections.Generic;
 using System.Linq;
-using System.Threading;
+using System.Timers;
 using System.Threading.Tasks;
+using System;
 
 namespace ServerSideBlazorApp.Pages
 {
     public partial class Customers
     {
         #region internals
-        private static readonly SemaphoreSlim isSearching = new SemaphoreSlim(1, 1);
         private static readonly Sort DefaultSort = PagingParameters.CreateDefaultSort(nameof(CustomerSummaryModel.Name), SortDirection.Ascending);
         private static readonly IList<int> AllowedPageSizes = new int[] { 5, 10, 25, 50, 100 };
-
-        private string SearchPhrase { get; set; }
-        private PagingParameters PagingParameters { get; set; } = PagingParameters.CreateDefault(DefaultSort);
-        private Page<CustomerSummaryModel> CurrentPage { get; set; } = Page<CustomerSummaryModel>.CreateDefault();
-        private PagingParameters PreviousPagingParameters { get; set; }
-        #endregion
+        private readonly Timer searchTimer = new(800);
+        private bool isFromReturnUrl;
+        private bool ShowProgressBar { get; set; } = false;
+        private Page<CustomerSummaryModel, PagingParametersWithSearch> CurrentPage { get; set; } = new Page<CustomerSummaryModel, PagingParametersWithSearch>(PagingParametersWithSearch.CreateDefault(DefaultSort), null, 0);
+        #endregion 
 
         #region methods
         private async Task FetchCurrentPageAsync()
         {
-            await ProgressBar.Show();
+            ShowProgressBar = true;
 
             try
             {
-                CurrentPage = await service.GetSummaryPageAsync(PagingParameters, SearchPhrase);
-                PreviousPagingParameters = PagingParameters;
+                var page = await CustomerService.GetSummaryPageAsync(CurrentPage.PagingParameters);
+                CurrentPage.Data = page.Data;
+                CurrentPage.TotalCount = page.TotalCount;
             }
             finally
             {
-                await ProgressBar.Hide();
+                ShowProgressBar = false;
             }
-
-            StateHasChanged();
         }
 
-        private async Task OnPage(DataGridReadDataEventArgs<CustomerSummaryModel> args)
+        private async Task OnReadData(DataGridReadDataEventArgs<CustomerSummaryModel> args)
         {
             if (!args.Columns.Any())
                 return;
-
-            PagingParameters = args.CreatePageRequestModel(PreviousPagingParameters ?? PagingParameters, DefaultSort);
-
-            await FetchCurrentPageAsync();
+            var requested = args.CreatePagingParameters(CurrentPage.PagingParameters, DefaultSort, p =>
+            {
+                p.SearchPhrase = CurrentPage.PagingParameters.SearchPhrase;
+                if (isFromReturnUrl)
+                {
+                    if (p.Offset == CurrentPage.PagingParameters.Offset)
+                        isFromReturnUrl = false; //offset same as from  returnUrl or is now in-sync
+                    else
+                        p.Offset = CurrentPage.PagingParameters.Offset;  //from returnUrl, the parameter from returnUrl should override
+                }
+            });
+            if (CurrentPage.PagingParameters != requested)
+            {
+                CurrentPage.PagingParameters = requested;
+                CurrentPage.Data = null;
+            }
+            if (CurrentPage.Data is null)
+                await FetchCurrentPageAsync();
         }
 
-        private async Task OnSearch(string searchPhrase)
+        private void OnPageChanged(DataGridPageChangedEventArgs args)
         {
-            if (SearchPhrase == searchPhrase)
-                return;            
+            var requested = args.CreatePagingParameters(CurrentPage.PagingParameters, DefaultSort, p => p.SearchPhrase = CurrentPage.PagingParameters.SearchPhrase);
+            if (CurrentPage.PagingParameters != requested)
+            {
+                CurrentPage.PagingParameters = requested;
+                CurrentPage.Data = null;
+            }
+        }
 
-            await isSearching.WaitAsync();
-            try
+        private void OnPageSizeChanged(int pageSize)
+        {
+            if (pageSize != CurrentPage.PagingParameters.Limit)
             {
-                SearchPhrase = searchPhrase;
-                PagingParameters.Offset = 0;
-                await FetchCurrentPageAsync();
+                CurrentPage.PagingParameters.Limit = pageSize;
+                CurrentPage.PagingParameters.Offset = 0;
+                CurrentPage.Data = null;
             }
-            finally
+        }
+
+        private void OnSearch(string searchPhrase)
+        {
+            CurrentPage.PagingParameters.SearchPhrase = searchPhrase;
+            CurrentPage.Data = null;
+            searchTimer.Stop();
+            searchTimer.Start();
+        }
+
+        private void OnSearch(object source, ElapsedEventArgs e)
+        {
+            InvokeAsync(async () =>
             {
-                isSearching.Release();
-            }
+                CurrentPage.PagingParameters.Offset = 0;
+                if (CurrentPage.PagingParameters.Offset == 0)
+                {
+                    CurrentPage.Data = null;
+                    await FetchCurrentPageAsync();
+                    StateHasChanged();
+                }
+            });
         }
 
         private string BuildDetailUrl(int id)
-            => $"/customers/{id}?{NavigationManager.ToReturnUrl("customers", PagingParameters)}";
+        {
+            var url = $"/customers/{id}?{NavigationManager.ToReturnUrl("customers", CurrentPage.PagingParameters)}";
+            return url;
+        }
+
+        protected override void OnInitialized()
+        {
+            searchTimer.Elapsed += OnSearch;
+            searchTimer.AutoReset = false;
+            base.OnInitialized();
+        }
 
         public async override Task SetParametersAsync(ParameterView parameters)
         {
-            if (NavigationManager.TryGetPagingParametersFromReturnUrl(out PagingParameters page))
-                PagingParameters = page;
+            if (NavigationManager.TryGetPagingParametersFromReturnUrl(out PagingParametersWithSearch page))
+            {
+                CurrentPage.PagingParameters = page;
+                isFromReturnUrl = true;
+            }
+
             await base.SetParametersAsync(parameters);
+        }
+
+        void IDisposable.Dispose()
+        {
+            searchTimer?.Dispose();
         }
         #endregion
     }