+++ samples/mssql/ServerSideBlazorApp/Shared/CustomerOrdersView.razor.cs
@@ -32,7 +32,7 @@ namespace ServerSideBlazorApp.Shared
             if (!args.Columns.Any())
                 return;
 
-            PagingParameters = args.CreatePageRequestModel(PreviousPagingParameters ?? PagingParameters, DefaultSort);
+            PagingParameters = args.CreatePagingParameters(PreviousPagingParameters ?? PagingParameters, DefaultSort);
 
             await FetchCurrentPageAsync();
         }