+++ src/HatTrick.DbEx.Sql/Configuration/_SqlStatements/_Execution/IExecutionPipelineEventConfigurationBuilder.cs
@@ -91,6 +91,78 @@ namespace HatTrick.DbEx.Sql.Configuration
         /// </summary>
         /// <param name="predicate">A predicate used to determine if the <paramref name="action"/> is executed.</param>
         IExecutionPipelineEventConfigurationBuilder OnAfterSqlStatementAssembly(Func<AfterAssemblyPipelineExecutionContext, CancellationToken, Task> action, Predicate<AfterAssemblyPipelineExecutionContext> predicate);
+
+        #region update
+        /// <summary>
+        /// Execute the delegate <paramref name="action"/> prior to assembling a sql statement from a query expression for an UPDATE operation.
+        /// </summary>
+        IExecutionPipelineEventConfigurationBuilder OnBeforeUpdateSqlStatementAssembly(Action<BeforeUpdateAssemblyPipelineExecutionContext> action);
+
+        /// <summary>
+        /// Execute the delegate <paramref name="action"/> prior to assembling a sql statement from a query expression for an UPDATE operation.
+        /// </summary>
+        /// <param name="predicate">A predicate used to determine if the <paramref name="action"/> is executed.</param>
+        IExecutionPipelineEventConfigurationBuilder OnBeforeUpdateSqlStatementAssembly(Action<BeforeUpdateAssemblyPipelineExecutionContext> action, Predicate<BeforeUpdateAssemblyPipelineExecutionContext> predicate);
+
+        /// <summary>
+        /// Execute the delegate <paramref name="action"/> prior to assembling a sql statement from a query expression for an UPDATE operation.
+        /// </summary>
+        IExecutionPipelineEventConfigurationBuilder OnBeforeUpdateSqlStatementAssembly(Func<BeforeUpdateAssemblyPipelineExecutionContext, Task> action);
+
+        /// <summary>
+        /// Execute the delegate <paramref name="action"/> prior to assembling a sql statement from a query expression for an UPDATE operation.
+        /// </summary>
+        /// <param name="predicate">A predicate used to determine if the <paramref name="action"/> is executed.</param>
+        IExecutionPipelineEventConfigurationBuilder OnBeforeUpdateSqlStatementAssembly(Func<BeforeUpdateAssemblyPipelineExecutionContext, Task> action, Predicate<BeforeUpdateAssemblyPipelineExecutionContext> predicate);
+
+        /// <summary>
+        /// Execute the delegate <paramref name="action"/> prior to assembling a sql statement from a query expression for an UPDATE operation.
+        /// </summary>
+        IExecutionPipelineEventConfigurationBuilder OnBeforeUpdateSqlStatementAssembly(Func<BeforeUpdateAssemblyPipelineExecutionContext, CancellationToken, Task> action);
+
+        /// <summary>
+        /// Execute the delegate <paramref name="action"/> prior to assembling a sql statement from a query expression for an UPDATE operation.
+        /// </summary>
+        /// <param name="predicate">A predicate used to determine if the <paramref name="action"/> is executed.</param>
+        IExecutionPipelineEventConfigurationBuilder OnBeforeUpdateSqlStatementAssembly(Func<BeforeUpdateAssemblyPipelineExecutionContext, CancellationToken, Task> action, Predicate<BeforeUpdateAssemblyPipelineExecutionContext> predicate);
+
+        #endregion
+
+        #region insert
+        /// <summary>
+        /// Execute the delegate <paramref name="action"/> prior to assembling a sql statement from a query expression for an UPDATE operation.
+        /// </summary>
+        IExecutionPipelineEventConfigurationBuilder OnBeforeInsertSqlStatementAssembly(Action<BeforeInsertAssemblyPipelineExecutionContext> action);
+
+        /// <summary>
+        /// Execute the delegate <paramref name="action"/> prior to assembling a sql statement from a query expression for an UPDATE operation.
+        /// </summary>
+        /// <param name="predicate">A predicate used to determine if the <paramref name="action"/> is executed.</param>
+        IExecutionPipelineEventConfigurationBuilder OnBeforeInsertSqlStatementAssembly(Action<BeforeInsertAssemblyPipelineExecutionContext> action, Predicate<BeforeInsertAssemblyPipelineExecutionContext> predicate);
+
+        /// <summary>
+        /// Execute the delegate <paramref name="action"/> prior to assembling a sql statement from a query expression for an UPDATE operation.
+        /// </summary>
+        IExecutionPipelineEventConfigurationBuilder OnBeforeInsertSqlStatementAssembly(Func<BeforeInsertAssemblyPipelineExecutionContext, Task> action);
+
+        /// <summary>
+        /// Execute the delegate <paramref name="action"/> prior to assembling a sql statement from a query expression for an UPDATE operation.
+        /// </summary>
+        /// <param name="predicate">A predicate used to determine if the <paramref name="action"/> is executed.</param>
+        IExecutionPipelineEventConfigurationBuilder OnBeforeInsertSqlStatementAssembly(Func<BeforeInsertAssemblyPipelineExecutionContext, Task> action, Predicate<BeforeInsertAssemblyPipelineExecutionContext> predicate);
+
+        /// <summary>
+        /// Execute the delegate <paramref name="action"/> prior to assembling a sql statement from a query expression for an UPDATE operation.
+        /// </summary>
+        IExecutionPipelineEventConfigurationBuilder OnBeforeInsertSqlStatementAssembly(Func<BeforeInsertAssemblyPipelineExecutionContext, CancellationToken, Task> action);
+
+        /// <summary>
+        /// Execute the delegate <paramref name="action"/> prior to assembling a sql statement from a query expression for an UPDATE operation.
+        /// </summary>
+        /// <param name="predicate">A predicate used to determine if the <paramref name="action"/> is executed.</param>
+        IExecutionPipelineEventConfigurationBuilder OnBeforeInsertSqlStatementAssembly(Func<BeforeInsertAssemblyPipelineExecutionContext, CancellationToken, Task> action, Predicate<BeforeInsertAssemblyPipelineExecutionContext> predicate);
+
+        #endregion
         #endregion
 
         #region insert