+++ src/HatTrick.DbEx.Sql/Expression/_Field/NullableStringFieldExpression.cs
@@ -37,6 +37,11 @@ namespace HatTrick.DbEx.Sql.Expression
         public abstract NullableStringElement As(string alias);
         #endregion
 
+        #region like
+        public FilterExpressionSet Like(string phrase)
+            => new FilterExpressionSet(new FilterExpression(this, new LikeExpression(phrase), FilterExpressionOperator.None));
+        #endregion
+
         #region equals
         public bool Equals(NullableStringFieldExpression obj)
             => obj is NullableStringFieldExpression && base.Equals(obj);