+++ src/HatTrick.DbEx.Sql/Expression/_Function/_DataType/_Concat/StringConcatFunctionExpression.cs
@@ -40,6 +40,11 @@ namespace HatTrick.DbEx.Sql.Expression
             => new StringSelectExpression(this).As(alias);
         #endregion
 
+        #region like
+        public FilterExpressionSet Like(string phrase)
+            => new FilterExpressionSet(new FilterExpression(this, new LikeExpression(phrase), FilterExpressionOperator.None));
+        #endregion
+
         #region equals
         public bool Equals(StringConcatFunctionExpression obj)
             => obj is StringConcatFunctionExpression && base.Equals(obj);