+++ src/HatTrick.DbEx.Sql/Expression/_Mediator/NullableStringExpressionMediator.cs
@@ -45,6 +45,11 @@ namespace HatTrick.DbEx.Sql.Expression
             => new NullableStringSelectExpression(this).As(alias);
         #endregion
 
+        #region like
+        public FilterExpressionSet Like(string phrase)
+            => new FilterExpressionSet(new FilterExpression(this, new LikeExpression(phrase), FilterExpressionOperator.None));
+        #endregion
+
         #region equals
         public bool Equals(NullableStringExpressionMediator obj)
             => obj is NullableStringExpressionMediator && base.Equals(obj);